+++ src/cmsc420/schema/spatial/SpatialStructure.java
@@ -6,6 +6,6 @@ import cmsc420.schema.DataStructure;
 public interface SpatialStructure extends DataStructure<City> {
 	
 	public int size();
-	public int getSpatialWidth();
-	public int getSpatialHeight();
+	public float getSpatialWidth();
+	public float getSpatialHeight();
 }
